Continue Restart Device Work
almost ready for testing

+++ Structs/SP_DEVINFO_DATA.cs
@@ -7,6 +7,6 @@ public struct SP_DEVINFO_DATA
 {
     public UInt32 cbSize;
     public Guid ClassGuid;
-    public UInt32 DevInst;
+    public IntPtr DevInst;
     public IntPtr Reserved;
 }
